Add files via upload
I made a number of changes to make the document more consistent and totally text based to facilitate using it to populate data fields.
it's probably good to keep the original with the symbols for human reference.  This should be the version for the project's use.

-CSV format, Excel format is available.
-Reformatted font and style to vanilla minimalist.
-Fixed capitalization and spelling in a few places for consistency.
-Made tree, ability, modifier titles with 2 words like "Wind Caller" upper case on both for cleaner in game display.
-Removed row 143, condensed it since the filled fields referred to the tree above.
-Added Favored Move to all the baby trees for completeness.

This should be 99% complete, didn't triple check and it's possible I lost some minor edits from dealing with multiple versions of this.
Placing this here for ease of access as well as to increase the girth of my contribution bar.