Found pesky error, streamlined Autoselect.
AutoSelect is now done via the Generate button and dropdown menu.
Filling a tree will no longer throw an error if you roll on it, and 99% of errors are gone.

Levels have no cap, but the trees will be full after L48. Currently anything above L40 should require a complete revamp to allow tree switching, so I've squashed the bugs on the lower levels. It's possible to lose a couple levels from full trees if you build a BL 8 pokemon; there should only be 2 empty boxes but sometimes there can be more.

Code is cleaned up and some of the Autoselect functionality was removed and replaced. There is now a master list of all bonuses gained, to facilitate switching trees and keeping records.

We need to retool how we call rolls, as well as adding the ability to switch trees. Currently there are empty lists of bonuses at the higher levels, I can't figure out how to roll only on certain trees.