Started reworking level up system
Now includes maturity, rate, and a script to calculate them at level up. Currently only works for "Generate" button but could be added to manual selection. Has no UI yet, but lists the bonuses and keeps track of the highest maturity reached so we don't get bonuses twice. We currently delete items as they are gained, but they can be saved to another list if needed.

A lot more is being added on to the new scripts and may need to be refactored, CallTreeRoll and TreeManager will also need to be rebuilt due to the increased complexity, so they can handle active trees and different tiers. Currently rolling on all 4 trees is required and 11/12 don't work as intended.

Also adds a shiny checker at full odds for Gen 6 and later.