PokeSheetDisplay
Pokemon now displays on the character sheet with most parameters being assigned manually at GameManager.Awake.

Stats calculate and display properly, BG colors assign based on breed and text input reassigns Nickname and Trainer for the pokemon.

Still need to call updates to display fields on level up, among other things so that the text reflects the changes, and be able to swap in different pokemon.