Fixed crashing issue.
Cause was caused because the constructor was being called inside the constructor causing an inflinite loop. This was hard to notice because it was occurring inside an implicit operator that was converting an int into the corresponding type(by calling the constructor). I impliminted the minimum neccessary solution of adding a constructor in the generic constructor  (in PokemonStatGeneric<T>) that takes an int instead of a stat. This stoped the implicit conversion from being called.